Phase 2: Single-Chain Configuration and ABI Consolidation
- Consolidated ABIs by moving morpho_abi.ts to packages/contracts/abi
- Deleted duplicate apps/web/abi directory
- Updated all ABI imports to use packages/contracts/abi
- Simplified wagmi.ts to support only Katana network
- Removed NetworkSelector component and all references
- Updated Auth.tsx to focus on single-chain Katana experience
- Fixed FlowSwapModal and related imports to use app/lib/networks
- Build completes successfully with no compilation errors

BREAKING CHANGES:
- Application now only supports Katana network (chain ID: 747474)
- Removed multi-chain support for Saga and Arbitrum
- NetworkSelector UI component removed
